requirements.txt
python-telegram-bot==20.7
psycopg2-binary
streamlit
pandas